Lesson18 update

+++ delete.php
@@ -1,9 +1,17 @@
 <?php
 session_start();
 
+ini_set('error_reporting', E_ALL);
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+
 $id = $_GET['id'];
 $name = $_GET['name'];
-unlink("img/new/" . $name);
+
+if (file_exists("img/new/" . $name)) {
+    unlink("img/new/" . $name);
+}
+
 
 $pdo = new PDO("mysql:host=localhost;dbname=php_course;", "root", "");
 $sql = "DELETE FROM images WHERE id=:id";
